requirements.txt
Jinja2==3.1.2
pyodbc==5.0.1
Requests==2.32.3
xhtml2pdf==0.2.16
Flask==3.0.3
pandas==2.2.2
APScheduler==3.10.4
python-telegram-bot==21.5
python-dotenv == 1.0.1
